Add the required change of reviewer
