+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "homepage": "https://cynthiainga.github.io/bookstore/",
   "dependencies": {
+    "@reduxjs/toolkit": "^1.8.0",
     "@testing-library/jest-dom": "^5.16.3",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
