+++ src/components/categoriesRoute.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const Categories = () => {
+  const styling = {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+  };
+
+  return (
+    <div style={styling}>
+      <h1>Under Construction</h1>
+    </div>
+  );
+};
+
+export default Categories;
