+++ src/redux/configureStore.js
@@ -1,9 +1,11 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
 import booksReducer from './books/books';
+import categoriesReducer from './categories/categories';
 
 const rootReducer = combineReducers({
   booksReducer,
+  categoriesReducer,
 });
 
 const store = createStore(rootReducer, applyMiddleware(logger));
